feat: add ignored rom dirs (#29)
This pull request includes several changes to the configuration and
scanning logic for the ROM management system. The most notable updates
involve the addition of a feature to ignore specific directories during
the scan, the usage of the `slices` package for sorting, and some
refactoring for better readability and maintainability.

### Configuration updates:
* Added `IgnoreDirs` field to the `userConfigs` struct and the
corresponding initialization in `config/config.go`.
[[1]](diffhunk://#diff-fe44f09c4d5977b5f5eaea29170b6a0748819c9d02271746a20d81a5f3efca17R63)
[[2]](diffhunk://#diff-fe44f09c4d5977b5f5eaea29170b6a0748819c9d02271746a20d81a5f3efca17R113)
[[3]](diffhunk://#diff-fe44f09c4d5977b5f5eaea29170b6a0748819c9d02271746a20d81a5f3efca17R133)
* Updated `includes/screech.yaml` to include a list of directories to
ignore during the scan.

### Codebase improvements:
* Replaced manual sorting logic with the `slices.SortFunc` utility for
better performance and readability in `screens/home.go`.
* Refactored the directory filtering logic into separate functions
`filterDirs`, `hasVisibleEntries`, and `isAllowedDir` to improve code
clarity in `screens/home.go`.
[[1]](diffhunk://#diff-27acbaff362aa484890bd6d494b8fa77f3423d0bc53395dc8c292f61ad5c8132L176-R187)
[[2]](diffhunk://#diff-27acbaff362aa484890bd6d494b8fa77f3423d0bc53395dc8c292f61ad5c8132R198-R220)

### Minor changes:
* Modified the event message to trim the base directory path in
`screens/scraping.go`.
* Simplified the worker pool creation loop in `screens/scraping.go`.